firmware/requirements.txt
Adafruit-Blinka==8.19.0
adafruit-blinka-bleio==4.1.0
adafruit-circuitpython-ble==10.0.3
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==1.14.0
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.46.0
Adafruit-PureIO==1.1.11
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
async-timeout==4.0.2
beautifulsoup4==4.12.2
bleak==0.20.2
branca==0.6.0
certifi==2022.12.7
cftime==1.6.2
charset-normalizer==2.1.1
conda==4.3.16
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
docopt==0.6.2
folium==0.14.0
fonttools==4.39.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
getmac==0.8.3
heartpy==1.2.7
idna==3.4
importlib-resources==5.12.0
influxdb-client==1.36.1
install==1.3.5
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.2
matplotlib==3.7.1
mplleaflet==0.0.5
netifaces==0.11.0
network==0.1
numba==0.56.4
numexpr==2.8.6
numpy==1.24.2
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.0
Pillow==9.5.0
plotly==5.15.0
pycosat==0.6.3
pydap==3.4.0
pyftdi==0.54.0
pynmea2==1.18.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
pyparsing==3.0.9
pyproj==3.5.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
pyusb==1.2.1
PyYAML==5.4.1
reactivex==4.0.4
requests==2.28.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
seabreeze==2.1.0
shapely==2.0.0
shapley==1.0.3
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sklearn==0.0
soupsieve==2.4.1
tables==3.7.0
tenacity==8.2.2
threadpoolctl==3.1.0
typing_extensions==4.0.1
tzdata==2023.3
urllib3==1.26.13
WebOb==1.8.7
xarray==2023.4.0
zipp==3.15.0
